Minor updates to thoughts
Added a GetThought method to the container classes of the thought strings to return a random entry in the arrays. I'm sure this can be refactored to remove duplication, but I'm just trying to get something working for the moment. Also added a new click UI priority for the thoughts UI and the ThoughtsUI class